pyproject.toml
[project]
name = "redpen"
version = "0.1.0"
description = "A Drake Equation for English writing quality - CLI tool for prose analysis"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Andrew Brookins", email = "a@andrewbrookins.com" }]
keywords = ["writing", "grammar", "style", "lint", "prose", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "textstat>=0.7",
    "proselint>=0.14",
    "pyspellchecker>=0.8",
    "litellm>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "toml>=0.10",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.0",
    "ruff>=0.4",
]
languagetool = [
    "language-tool-python>=2.8",
]

[project.scripts]
redpen = "redpen.cli.main:cli"

[project.urls]
Homepage = "https://github.com/abrookins/redpen"
Repository = "https://github.com/abrookins/redpen"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/redpen"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

